Buffer / data version counters: Long -> Int
The version counters used to gate buffer uploads — AbstractShape.bufferDataVersion,
BufferObject.version, UploadQueue.queueBufferUpload's parameter, UploadQueue.Entry.version,
RenderContext.offerGLBufferUpload's parameter, ColladaScene.bufferVersion, GltfScene.bufferVersion,
StarFieldLayer.startDataVersion / sunDataVersion — were all Long. Int max (~2.1B) is years of
runtime even at hundreds of mutations per second, so Long was overkill. Trimmed to Int across
the chain: 4 bytes saved per shape, per buffer object, per upload queue entry, per data version
counter — and Int.hashCode is a no-op identity vs Long's 64->32 bit fold.

+++ worldwind/src/commonMain/kotlin/earth/worldwind/draw/UploadQueue.kt
@@ -15,7 +15,7 @@ open class UploadQueue internal constructor() {
         protected const val MIN_CAPACITY_INCREMENT = 12
     }
 
-    fun queueBufferUpload(buffer: BufferObject, array: NumericArray, version: Long) {
+    fun queueBufferUpload(buffer: BufferObject, array: NumericArray, version: Int) {
         val capacity = entries.size
         if (capacity == size) {
             val increment = max(capacity shr 1, MIN_CAPACITY_INCREMENT)
@@ -60,12 +60,12 @@ open class UploadQueue internal constructor() {
     protected open class Entry {
         var array: NumericArray? = null
         var buffer: BufferObject? = null
-        var version = 0L
+        var version = 0
 
         fun recycle() {
             array = null
             buffer = null
-            version = 0L
+            version = 0
         }
     }
 }
