+++ worldwind/src/commonMain/kotlin/earth/worldwind/render/RenderContext.kt
@@ -349,7 +349,7 @@ open class RenderContext {
         } else null
     }
 
-    inline fun offerGLBufferUpload(key: Any, newVersion: Long, arrayBuilder: () -> NumericArray) {
+    inline fun offerGLBufferUpload(key: Any, newVersion: Int, arrayBuilder: () -> NumericArray) {
         (renderResourceCache[key] as? BufferObject)?.let { buffer ->
             if (buffer.version < newVersion) {
                 val array = arrayBuilder()
